new22

+++ src/main/java/bHibernate/lesson4/E_DAO/OrderDAO.java
@@ -1,20 +1,15 @@
 package bHibernate.lesson4.E_DAO;
 
-import bHibernate.lesson4.B_model.ENUM.UserType;
-import bHibernate.lesson4.B_model.Hotel;
 import bHibernate.lesson4.B_model.Order;
 import bHibernate.lesson4.B_model.Room;
 import bHibernate.lesson4.B_model.User;
-import bHibernate.lesson4.C_controller.OrderController;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
 import org.hibernate.query.NativeQuery;
-
 import javax.persistence.NoResultException;
 import java.util.Date;
-import java.util.List;
 
 public class OrderDAO {
     private static SessionFactory sessionFactory = UserDAO.createSessionFactory();
@@ -69,7 +64,7 @@ public class OrderDAO {
             session = sessionFactory.openSession();
             tr = session.getTransaction();
             tr.begin();
-                order = (Order) session.load(Order.class,id);
+            order = (Order) session.get(Order.class,id);
             tr.commit();
             System.out.println("Find By ID is done");
         } catch (HibernateException | NoResultException e) {
@@ -84,22 +79,17 @@ public class OrderDAO {
         return order;
     }
 
-    public static void delete(User user,Room room,Date dateFrom, Date dateTo){
+    public static void delete(long userId, long roomId){
         try {
             session = sessionFactory.openSession();
             tr = session.getTransaction();
             tr.begin();
-                 Order order=creatObject(user,room,dateFrom,dateTo,0);
-                 session.delete(order);
-;
-//                session.delete(hotel);
-//                Order hotel = session.load(Hotel.class,50);
-//                if (hotel != null) {
-//                    session.delete(hotel);
-//                }
+                Order order = (Order) session.get(Order.class,sqlQueriOrder(userId, roomId));
+                if (order != null)
+                    session.delete(order);
             tr.commit();
             System.out.println("Delete is done");
-        } catch (HibernateException e) {
+        } catch (HibernateException | NoResultException  e) {
             System.err.println("Delete is failed");
             System.out.println(e.getMessage());
             if (tr != null)
@@ -109,15 +99,6 @@ public class OrderDAO {
                 session.close();
         }
     }
-
-
-
-    public static void main(String[] args) {
-//        OrderDAO.delete(UserDAO.findById(17), RoomDAO.findById(12), new Date(), new Date());
-
-    }
-
-
 
  //---------------------------------------------------------------------------------------------------------------
     private static Order creatObject(User user,Room room,Date dateFrom, Date dateTo, double moneyPaid) {
@@ -130,4 +111,13 @@ public class OrderDAO {
     return order;
     }
 
+    public static long sqlQueriOrder(long userId, long roomId){
+        NativeQuery nq = session.createNativeQuery(
+                "SELECT * FROM ORDERFPR WHERE USERFPR_ID = :param1 AND ROOMFPR_ID = :param2 ", Order.class);
+                        nq.setParameter("param1", userId);
+                        nq.setParameter("param2", roomId);
+        Order order= (Order) nq.getSingleResult();
+        System.out.println(order.getId());
+        return order.getId();
+    }
 }
